test(scale-up): move EC2 integration tests
